Split out controller module.
